requirements.txt
nltk
torch
transformers
FastAPI
python-dotenv
uvicorn